Change the default facility rank to 0 & regen
The simplest fix to the problem of land displaying a facility rank
of 1, even when there isn't a facility, is to set the default to
0 and manually set it for each new facility. The code is currently not
designed to support specific replacement rules, so this works
around the code. Admittedly, manually setting it is a bit annoying,
and could perhaps be replaced if the code is changed to support
easier extension, but this is a realistic depiction of what
is actually happening to the land.